move tests outside of python package
