+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 minversion = 6.0
 testpaths =
-    gnina/tests/
+    tests/
