+++ tests/test_inference.py
@@ -1,27 +1,6 @@
 import os
-import sys
-
-import pytest
 
 from gnina import inference, training
-
-
-@pytest.fixture
-def trainfile() -> str:
-    gnina_path = os.path.dirname(sys.modules["gnina"].__file__)
-    return os.path.join(gnina_path, "data", "test.types")
-
-
-@pytest.fixture
-def testfile() -> str:
-    gnina_path = os.path.dirname(sys.modules["gnina"].__file__)
-    return os.path.join(gnina_path, "data", "test.types")
-
-
-@pytest.fixture
-def dataroot() -> str:
-    gnina_path = os.path.dirname(sys.modules["gnina"].__file__)
-    return os.path.join(gnina_path, "data", "test")
 
 
 def test_inference(trainfile, testfile, dataroot, tmpdir, device):
