+++ tests/test_losses.py
