+++ tests/test_setup.py
@@ -1,22 +1,7 @@
-import os
-import sys
-
 import pytest
 import torch
 
 from gnina import setup, training
-
-
-@pytest.fixture
-def trainfile() -> str:
-    gnina_path = os.path.dirname(sys.modules["gnina"].__file__)
-    return os.path.join(gnina_path, "data", "test.types")
-
-
-@pytest.fixture
-def dataroot() -> str:
-    gnina_path = os.path.dirname(sys.modules["gnina"].__file__)
-    return os.path.join(gnina_path, "data", "test")
 
 
 def test_setup_example_provider_default(trainfile, dataroot, device):
